Tighten optimizeopt parity and raw-buffer arithmetic
Restore class-based OpRef::is_constant dispatch with debug_assert on
typed body variants carrying CONST_BIT; align make_nonnull early
return for Type::Int per optimizer.py; let plain_int_w panic on
_fits_int precondition violation instead of silently returning 0.

Wire SetarrayitemRaw to optimize_setarrayitem_raw, drop the dead
OptVirtualize SetarrayitemRaw vable mirror, and remove dead Pyre-only
raw buffer coverage helpers. Switch raw buffer offsets to i64 to
match rawbuffer.py, drop the _R fold path, and add checked overlap
plus pointer-array guards.

Make shortpreamble and raw buffer constant allocation
namespace-aware with checked size casts, fix the infinite Vec::resize
hang on CONST_BIT-tainted aliases, and admit fits_int W_LongObject
into the list int-strategy.

Drop the dead GenerateGuardState path from optimizeopt/virtualstate
including its fake OpRef::from_raw(10_000+idx) GUARD_VALUE
synthesis; production guard generation flows through
GuardRequirement::GuardValue with ctx.make_constant_*.

+++ majit/majit-backend/src/lib.rs
@@ -150,15 +150,16 @@ pub enum ExitVirtualLayout {
     },
     /// resume.py:717 VRawSliceInfo — base_buffer + offset.
     RawSlice {
-        offset: usize,
+        /// info.py:460 signed slice base.
+        offset: i64,
         base: ExitValueSourceLayout,
     },
     RawBuffer {
         /// resume.py:694: self.func
         func: i64,
         size: usize,
-        /// resume.py:695: self.offsets
-        offsets: Vec<usize>,
+        /// resume.py:695: self.offsets — signed (rawbuffer.py:14).
+        offsets: Vec<i64>,
         /// resume.py:697: self.descrs
         descrs: Vec<majit_ir::ArrayDescrInfo>,
         /// resume.py:693: fieldnums (decoded)
